Relocate ProviderConfigSchemaSource
Relocate the ProviderConfigSchemaSource type to environs removing
more items from state.
