+++ apiserver/facades/client/client/client.go
@@ -21,6 +21,7 @@ import (
 	"github.com/juju/juju/core/multiwatcher"
 	coreos "github.com/juju/juju/core/os"
 	"github.com/juju/juju/core/permission"
+	"github.com/juju/juju/environs"
 	envtools "github.com/juju/juju/environs/tools"
 	"github.com/juju/juju/internal/cloudconfig/podcfg"
 	"github.com/juju/juju/internal/docker"
@@ -160,7 +161,7 @@ func NewFacade(ctx facade.ModelContext) (*Client, error) {
 			State:                    st,
 			model:                    model,
 			session:                  nil,
-			configSchemaSourceGetter: stateenvirons.ProviderConfigSchemaSource(serviceFactory.Cloud()),
+			configSchemaSourceGetter: environs.ProviderConfigSchemaSource(serviceFactory.Cloud()),
 		},
 		&poolShim{pool: ctx.StatePool()},
 		storageAccessor,
